Schueler permissions

+++ apps/server/modules/schueler/api.ts
@@ -20,12 +20,28 @@ router.post('/', async (req: Request<{}, {}, CreateSchuelerRequestBody>, res): P
     res.status(200).json(msg);
 });
 
-router.get('/', async (req, res) => {
+router.get('/', async (req, res): Promise<any> => {
+
+    if(req.userId == undefined) {
+        return getNoSessionResponse(res);
+    }
+    if (!req.permissions?.[Berechtigung.SchuelerRead]) {
+        return getNoPermissionResponse(res)
+    }
+
     const msg = await getSchuelerStore().getSchueler()
     res.status(200).json(msg);
 });
 
-router.put('/:schuelerId', async (req: Request<{ schuelerId: string }, {}, EditSchuelerRequestBody>, res) => {
+router.put('/:schuelerId', async (req: Request<{ schuelerId: string }, {}, EditSchuelerRequestBody>, res): Promise<any> => {
+    
+    if(req.userId == undefined) {
+        return getNoSessionResponse(res);
+    }
+    if (!(req.permissions?.[Berechtigung.SchuelerUpdate] && req.permissions?.[Berechtigung.SchuelerRead])) {
+        return getNoPermissionResponse(res)
+    }
+
     const schueler = req.body
     const { schuelerId } = req.params
     schueler.id = parseInt(schuelerId);
@@ -34,14 +50,30 @@ router.put('/:schuelerId', async (req: Request<{ schuelerId: string }, {}, EditS
     res.status(msg.success ? 200 : 400).json(msg);
 });
 
-router.get('/:schuelerId', async (req, res) => {
+router.get('/:schuelerId', async (req, res): Promise<any> => {
+
+    if(req.userId == undefined) {
+        return getNoSessionResponse(res);
+    }
+    if (!req.permissions?.[Berechtigung.SchuelerRead]) {
+        return getNoPermissionResponse(res)
+    }
+
     const { schuelerId } = req.params
     const msg = await getSchuelerStore().getSchuelerComplete(parseInt(schuelerId))
     console.log(msg)
     res.status(200).json(msg);
 });
 
-router.delete('/', async (req: Request<{}, {}, DeleteSchuelerRequestBody>, res: Response<DeleteSchuelerResponseBody>) => {
+router.delete('/', async (req: Request<{}, {}, DeleteSchuelerRequestBody>, res: Response<DeleteSchuelerResponseBody>): Promise<any> => {
+    
+    if(req.userId == undefined) {
+        return getNoSessionResponse(res);
+    }
+    if (!(req.permissions?.[Berechtigung.SchuelerUpdate] && req.permissions?.[Berechtigung.SchuelerRead])) {
+        return getNoPermissionResponse(res)
+    }
+    
     const schuelerId = req.body.schuelerId
     const msg = await getSchuelerStore().deleteSchueler(schuelerId)
     res.status(200).json(msg as DeleteSchuelerResponseBody);
